Turned AR Tracking into a utility and started outline detection
